ml_server/requirements.txt
tensorflow-macos
tensorflow-metal
flask
psycopg2-binary
scikit-learn